refactor: move prompt-cache into provider
